+++ pysteamauth/auth/__init__.py
@@ -1,14 +1,6 @@
-from .schemas import (
-    AuthenticatorData,
-    ServerTime,
-    ServerTimeResponse,
-)
 from .steam import Steam
 
 
 __all__ = [
     'Steam',
-    'AuthenticatorData',
-    'ServerTime',
-    'ServerTimeResponse',
 ]