+++ pysteamauth/auth/steam.py
@@ -2,18 +2,19 @@ import base64
 import binascii
 import hashlib
 import hmac
+import json
 import math
 from struct import pack
 from typing import (
     Any,
     Mapping,
     Optional,
-    Type,
 )
 
 import rsa
 from aiohttp import FormData
 from bitstring import BitArray
+from urllib3.util import parse_url
 
 from pysteamauth.abstract import (
     CookieStorageAbstract,
@@ -37,50 +38,62 @@ from pysteamauth.pb2.steammessages_auth.steamclient_pb2 import (
     EAuthTokenPlatformType,
 )
 
-from .schemas import (
-    AuthenticatorData,
-    FinalizeLoginStatus,
-    ServerTimeResponse,
-    SteamAuthorizationStatus,
-)
-from .utils import get_host_from_url
+from .schemas import FinalizeLoginStatus
 
 
 class Steam:
-
-    domains = (
-        'steamcommunity.com',
-        'store.steampowered.com',
-        'help.steampowered.com',
-    )
 
     def __init__(
         self,
         login: str,
         password: str,
-        steamid: int,
-        authenticator: Optional[AuthenticatorData] = None,
-        cookie_storage: Type[CookieStorageAbstract] = BaseCookieStorage,
-        request_strategy: Type[RequestStrategyAbstract] = BaseRequestStrategy,
+        steamid: Optional[int] = None,
+        shared_secret: Optional[str] = None,
+        identity_secret: Optional[str] = None,
+        device_id: Optional[str] = None,
+        cookie_storage: Optional[CookieStorageAbstract] = None,
+        request_strategy: Optional[RequestStrategyAbstract] = None,
     ):
         self._login = login
         self._steamid = steamid
         self._password = password
-        self._authenticator = authenticator
-        self._http = request_strategy()
-        self._storage = cookie_storage()
+        self._shared_secret = shared_secret
+        self._identity_secret = identity_secret
+        self._device_id = device_id
+        self._requests = request_strategy if request_strategy is not None else BaseRequestStrategy()
+        self._storage = cookie_storage if cookie_storage is not None else BaseCookieStorage()
 
     @property
     def steamid(self) -> int:
+        if not self._steamid:
+            raise ValueError('steamid is not specified')
         return self._steamid
 
     @property
-    def authenticator(self) -> Optional[AuthenticatorData]:
-        return self._authenticator
+    def shared_secret(self) -> str:
+        if not self._shared_secret:
+            raise ValueError('shared_secret is not specified')
+        return self._shared_secret
+
+    @property
+    def identity_secret(self) -> str:
+        if not self._identity_secret:
+            raise ValueError('identity_secret is not specified')
+        return self._identity_secret
+
+    @property
+    def device_id(self) -> str:
+        if not self._device_id:
+            raise ValueError('device_id is not specified')
+        return self._device_id
 
     @property
     def login(self) -> str:
         return self._login
+
+    @property
+    def partner_id(self) -> int:
+        return self.steamid - 76561197960265728
 
     async def cookies(self, domain: str = 'steamcommunity.com') -> Mapping[str, str]:
         return await self._storage.get(
@@ -92,20 +105,11 @@ class Steam:
         return (await self.cookies(domain))['sessionid']
 
     async def request(self, url: str, method: str = 'GET', **kwargs: Any) -> str:
-        """
-        Request with Steam session.
-        """
-        domain = get_host_from_url(url)
-
-        if domain not in self.domains:
-            raise RuntimeError(f'Wrong domain "{domain}"')
-
         cookies = await self._storage.get(
             login=self._login,
-            domain=domain,
+            domain=parse_url(url).host,
         )
-
-        return await self._http.text(
+        return await self._requests.text(
             url=url,
             method=method,
             cookies={
@@ -116,22 +120,24 @@ class Steam:
         )
 
     async def is_authorized(self) -> bool:
-        """
-        Is alive authorization.
-        """
         response: str = await self.request(
             url='https://steamcommunity.com/chat/clientjstoken',
         )
-        return SteamAuthorizationStatus.parse_raw(response).logged_in
+        data = json.loads(response)
+        logged_in = data['logged_in']
+        if logged_in:
+            if not self._steamid:
+                try:
+                    self._steamid = int(data['steamid'])
+                except KeyError:
+                    ...
+        return logged_in
 
     async def _getrsakey(self) -> CAuthentication_GetPasswordRSAPublicKey_Response:
-        """
-        Get rsa keys for password encryption.
-        """
         message = CAuthentication_GetPasswordRSAPublicKey_Request(
             account_name=self._login,
         )
-        response = await self._http.bytes(
+        response = await self._requests.bytes(
             method='GET',
             url='https://api.steampowered.com/IAuthenticationService/GetPasswordRSAPublicKey/v1',
             params={
@@ -145,9 +151,6 @@ class Steam:
             encrypted_password: str,
             rsa_timestamp: int,
     ) -> CAuthentication_BeginAuthSessionViaCredentials_Response:
-        """
-        Begin auth session.
-        """
         message = CAuthentication_BeginAuthSessionViaCredentials_Request(
             account_name=self._login,
             encrypted_password=encrypted_password,
@@ -158,7 +161,7 @@ class Steam:
             persistence=ESessionPersistence.k_ESessionPersistence_Persistent,
             device_friendly_name='Mozilla/5.0 (X11; Linux x86_64; rv:1.9.5.20) Gecko/2812-12-10 04:56:28 Firefox/3.8',
         )
-        response = await self._http.bytes(
+        response = await self._requests.bytes(
             method='POST',
             url='https://api.steampowered.com/IAuthenticationService/BeginAuthSessionViaCredentials/v1',
             data=FormData(
@@ -170,25 +173,18 @@ class Steam:
         return CAuthentication_BeginAuthSessionViaCredentials_Response.FromString(response)
 
     async def get_server_time(self) -> int:
-        """
-        Get server time.
-        """
-        response = await self._http.text(
+        response = await self._requests.text(
             method='POST',
             url='https://api.steampowered.com/ITwoFactorService/QueryTime/v0001',
         )
-        return ServerTimeResponse.parse_raw(response).response.server_time
+        data = json.loads(response)
+        return int(data['response']['server_time'])
 
-    async def get_steam_guard(self, server_time: int) -> str:
-        """
-        Calculating Steam Guard code.
-        """
-        if self._authenticator is None:
-            raise RuntimeError('Not found authenticator')
-
+    @classmethod
+    async def get_steam_guard(cls, shared_secret: str, server_time: int) -> str:
         data = binascii.unhexlify(
             hmac.new(
-                key=base64.b64decode(self._authenticator.shared_secret),
+                key=base64.b64decode(shared_secret),
                 msg=pack('!L', 0) + pack('!L', math.floor(server_time // 30)),
                 digestmod=hashlib.sha1,
             ).hexdigest(),
@@ -211,13 +207,8 @@ class Steam:
         return code
 
     def get_confirmation_hash(self, server_time: int, tag: str = 'conf') -> str:
-        """
-        Get mobile confirmation hash.
-        """
-        if self._authenticator is None:
-            raise RuntimeError('Not found authenticator')
         identitysecret = base64.b64decode(
-            self._authenticator.identity_secret,
+            self.identity_secret,
         )
         secret = BitArray(
             bytes=identitysecret,
@@ -238,9 +229,6 @@ class Steam:
         return str(base64.b64encode(confirmation.digest()), 'utf8')
 
     def _encrypt_password(self, keys: CAuthentication_GetPasswordRSAPublicKey_Response) -> str:
-        """
-        Encrypt password.
-        """
         publickey_exp = int(keys.publickey_exp, 16)  # type:ignore
         publickey_mod = int(keys.publickey_mod, 16)  # type:ignore
         public_key = rsa.PublicKey(
@@ -260,16 +248,13 @@ class Steam:
             code: str,
             code_type: int,
     ) -> None:
-        """
-        Update session request.
-        """
         message = CAuthentication_UpdateAuthSessionWithSteamGuardCode_Request(
             client_id=client_id,
             steamid=steamid,
             code=code,
             code_type=code_type,
         )
-        await self._http.request(
+        await self._requests.request(
             method='POST',
             url='https://api.steampowered.com/IAuthenticationService/UpdateAuthSessionWithSteamGuardCode/v1',
             data=FormData(
@@ -284,14 +269,11 @@ class Steam:
             client_id: int,
             request_id: bytes,
     ) -> CAuthentication_PollAuthSessionStatus_Response:
-        """
-        Auth session status.
-        """
         message = CAuthentication_PollAuthSessionStatus_Request(
             client_id=client_id,
             request_id=request_id,
         )
-        response = await self._http.bytes(
+        response = await self._requests.bytes(
             method='POST',
             url='https://api.steampowered.com/IAuthenticationService/PollAuthSessionStatus/v1',
             data=FormData(
@@ -303,10 +285,7 @@ class Steam:
         return CAuthentication_PollAuthSessionStatus_Response.FromString(response)
 
     async def _finalize_login(self, refresh_token: str, sessionid: str) -> FinalizeLoginStatus:
-        """
-        Finalize login.
-        """
-        response = await self._http.text(
+        response = await self._requests.text(
             method='POST',
             url='https://login.steampowered.com/jwt/finalizelogin',
             data=FormData(
@@ -320,10 +299,7 @@ class Steam:
         return FinalizeLoginStatus.parse_raw(response)
 
     async def _set_token(self, url: str, nonce: str, auth: str, steamid: int) -> None:
-        """
-        Set token.
-        """
-        await self._http.request(
+        await self._requests.request(
             method='POST',
             url=url,
             data=FormData(
@@ -336,19 +312,13 @@ class Steam:
         )
 
     def _is_twofactor_required(self, confirmation: CAuthentication_AllowedConfirmation) -> bool:
-        """
-        Is twofactor required.
-        """
         return confirmation.confirmation_type == EAuthSessionGuardType.k_EAuthSessionGuardType_DeviceCode
 
     async def login_to_steam(self) -> None:
-        """
-        Login to Steam.
-        """
         if await self.is_authorized():
             return
-        if not self._http.cookies().get('sessionid'):
-            await self._http.bytes(
+        if not self._requests.cookies().get('sessionid'):
+            await self._requests.bytes(
                 method='GET',
                 url='https://steamcommunity.com',
             )
@@ -361,7 +331,7 @@ class Steam:
         if auth_session.allowed_confirmations:
             if self._is_twofactor_required(auth_session.allowed_confirmations[0]):
                 server_time = await self.get_server_time()
-                code = await self.get_steam_guard(server_time)
+                code = await Steam.get_steam_guard(self.shared_secret, server_time)
                 await self._update_auth_session(
                     client_id=auth_session.client_id,
                     steamid=auth_session.steamid,
@@ -374,9 +344,8 @@ class Steam:
         )
         tokens = await self._finalize_login(
             refresh_token=session.refresh_token,
-            sessionid=self._http.cookies()['sessionid'],
+            sessionid=self._requests.cookies()['sessionid'],
         )
-
         for token in tokens.transfer_info:
             await self._set_token(
                 url=token.url,
@@ -384,15 +353,14 @@ class Steam:
                 auth=token.params.auth,
                 steamid=auth_session.steamid,
             )
-        # Needs to get additional cookies through aiohttp session
-        for domain in ('store', 'help'):
-            await self._http.bytes(
-                method='GET',
-                url=f'https://{domain}.steampowered.com',
-            )
-        # Write cookies from aiohttp session
-        cookies = {}
-        for domain in self.domains:
-            cookies[domain] = self._http.cookies(domain)
-        # Save cookies to storage
+        if not self._steamid and tokens.steamID:
+            self._steamid = int(tokens.steamID)
+        cookies = {
+            'steamcommunity.com': self._requests.cookies('steamcommunity.com'),
+        }
+        for url in ('https://store.steampowered.com', 'https://help.steampowered.com'):
+            await self._requests.bytes(url, 'GET')
+            cookies.update({
+                parse_url(url).host: self._requests.cookies(parse_url(url).host),
+            })
         await self._storage.set(login=self._login, cookies=cookies)